+++ MK4duo/src/sdcard/sdfat/SdFatConfig.h
@@ -87,7 +87,7 @@
  * SPI init rate for SD initialization commands. Must be 5 (F_CPU/64)
  * or 6 (F_CPU/128).
  */
-#define SPI_SD_INIT_RATE 11
+#define SPI_SD_INIT_RATE 5
 
 /**
  * Set the SS pin high for hardware SPI.  If SS is chip select for another SPI